requirements.txt
beautifulsoup4==4.14.2
html5lib==1.1
requests==2.32.5
selenium==4.36.0
webdriver-manager==4.0.2
python-dotenv==1.0.1
